third_party/blink/web_tests/external/wpt/html/rendering/dimension-attributes-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <hr width="0"> mapping to <hr> width property
  assert_equals: expected "0px" but got "1px"
[FAIL] <hr width="0%"> mapping to <hr> width property
  assert_equals: expected "0%" but got "1px"
[FAIL] <hr width="0px"> mapping to <hr> width property
  assert_equals: expected "0px" but got "1px"
[FAIL] <hr width="-0"> mapping to <hr> width property
  assert_equals: expected "auto" but got "1px"
[FAIL] <hr width="-0%"> mapping to <hr> width property
  assert_equals: expected "auto" but got "1px"
[FAIL] <hr width="+0"> mapping to <hr> width property
  assert_equals: expected "auto" but got "1px"
[FAIL] <hr width="+0%"> mapping to <hr> width property
  assert_equals: expected "auto" but got "1px"
Harness: the test ran to completion.
